terraform_v2/services/alert-manager/requirements.txt
functions-framework==3.8.1
google-cloud-firestore==2.19.0